+++ controllers/front/dispatcher.php
@@ -22,7 +22,6 @@ class OystDispatcherModuleFrontController extends ModuleFrontController
         Route::addRoute('PUT', '/script-tag', 'ScriptTag', 'setUrl');
         Route::addRoute('PUT', '/customer/search', 'Customer', 'search');
         Route::addRoute('PUT', '/cart/{id}', 'Cart', 'updateCart');
-        Route::addRoute('PUT', '/order/create', 'Order', 'createOrderFromCart');
 
         $request = new CurrentRequest();
